requirements.txt
fastapi==0.81.0
fhir.resources==6.4.0
pydantic==1.8.2
pytest==7.1.2
pytest-dotenv==0.5.2
requests==2.28.1
uvicorn==0.18.3
orjson>=3.6.8
